requirements.txt

#django-suit

coverage
#django-coverage

# django dev - need migration support
https://www.djangoproject.com/download/1.7c3/tarball/
django_cas_sso

model_mommy

#django-grappelli
# need dev version for django 1.7 compatibility
#git+git://github.com/sehmaschine/django-grappelli.git@dev/2.6.x

django-easyfilters
django-vanilla-views
django-crispy-forms
django-bootstrap3

#mezzanine
#mezzanine-grappelli
#-e git+git://github.com/renyi/mezzanine-blocks#egg=mezzanine-blocks

django-toolbelt 
